Added the settings menu to the RegisterActivity.  Reusing the PositMain menu with the Tacker and Projects options disabled.    This makes it possible for a user to change the server before logging in or creating an account.
This fixes the following bug:  it allows a user to recover from a situation where the current server is unreachable.  This could happen if the user changes the server setting to one that doesn't exist or isn't currently online.   If the user cannot reach a server, it will be impossible to register the phone.

NOTE:  This should probably be refactored.  An improved solution might be a button in the "Register" view or in the Login view??