Patch submitted by edward.bassett@ualberta.net for issue #95.
Screen rotation causes an Activity to be restarted, so that it can load in any new configuration required.
RegisterActivity has two different sub-activity-ish screens depending on the "action" that the user choses (Create New Account, or Login), and this selection was being lost on restart when RegisterActivity was started from PositMain. (When SettingsActivity kicked it off, the choice was saved in a recoverable way because there the user made the choice already at the settings screen, *before* RegisterActivity started.)
So the key here was to make the user's in-Activity choice persist somehow. onRetainNonConfigurationInstance() can save that value, and then onGetLastNonConfigurationInstance() (called from within onCreate(...)) can load that value if it exists (and proceed normally if it doesn't).
But first we had to capture the value of that choice when it is made - in a way that we can pass to onRetainNonConfiguration() for retention and later retrieval. To that end, I introduced a private member: String mAction.

These changes fix the Issue, without causing any new ones when RegisterActivity is called from SettingsActivity.

(Partially transplanted from 6b791a9c6001acbba1d97ff31d4435dfc8c75a08)